chore: rename files in the gitlab module
