+++ plugins/scaffolder-backend-module-gitlab/src/actions/gitlabIssueCreate.test.ts
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-import { createMockActionContext } from '@backstage/plugin-scaffolder-node-test-utils';
-import { createGitlabIssueAction, IssueType } from './createGitlabIssueAction';
 import { ConfigReader } from '@backstage/core-app-api';
 import { ScmIntegrations } from '@backstage/integration';
+import { createMockActionContext } from '@backstage/plugin-scaffolder-node-test-utils';
+import { createGitlabIssueAction, IssueType } from './gitlabIssueCreate';
 
 const mockGitlabClient = {
   Issues: {
