+++ plugins/scaffolder-backend-module-gitlab/src/actions/gitlabProjectAccessTokenCreate.examples.test.ts
@@ -15,10 +15,10 @@
  */
 import { ConfigReader } from '@backstage/config';
 import { ScmIntegrations } from '@backstage/integration';
-import yaml from 'yaml';
-import { createGitlabProjectAccessTokenAction } from './createGitlabProjectAccessTokenAction'; // Adjust the import based on your project structure
-import { examples } from './createGitlabProjectAccessTokenAction.examples';
 import { createMockActionContext } from '@backstage/plugin-scaffolder-node-test-utils';
+import yaml from 'yaml';
+import { createGitlabProjectAccessTokenAction } from './gitlabProjectAccessTokenCreate'; // Adjust the import based on your project structure
+import { examples } from './gitlabProjectAccessTokenCreate.examples';
 
 import { DateTime } from 'luxon';
