+++ plugins/scaffolder-backend-module-gitlab/src/actions/gitlabProjectVariableCreate.examples.ts
